FIX: be forgiving of type instability in TwoButtonShutter status PV
https://github.com/bluesky/ophyd/pull/946 changed it so that the value
passed to the callabck is the str version of the value not the
integer.

This means that the conversion of int -> str we are doing in the
callbacks of TwoButtonShutter are redundant and failing.  This means
that we will not ever mark the status object as done (and tend to fail
with a miss leading error due to trying to set the shutter again on
the way out of the RE).